auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.28.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis[async]>=5.0.0",
    "sqlalchemy>=2.0.43",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.11",
    "tox>=4.29.0",
]


[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]
default_section = "THIRDPARTY"

[tool.mypy]
python_version = 3.11
strict = true
ignore_missing_imports = true

[tool.ruff]
# Global configuration
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "alembic>=1.16.5",
    "asgi-lifespan>=2.1.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.11",
]
